Backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
